requirements.txt
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
pydantic>=2.6.0

# Windows-only. Installed via requirements-windows.txt on Win; skipped on macOS/Linux.
# MetaTrader5>=5.0.45

# Dev
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.6.0
freezegun>=1.5.0
